Relocated API calls to container parent component
